chore(crypto): Move crypto lib interfaces inside mod raw_signature (#849)
