+++ internal/crypto/src/raw_signature/openssl/cert_chain.rs
