+++ internal/crypto/src/raw_signature/openssl/signers/ecdsa_signer.rs
@@ -12,35 +12,60 @@
 // each license.
 
 use openssl::{
+    ec::EcKey,
+    hash::MessageDigest,
     pkey::{PKey, Private},
     sign::Signer,
     x509::X509,
 };
 
 use crate::{
-    openssl::{cert_chain::check_chain_order, OpenSslMutex},
-    raw_signature::{RawSigner, RawSignerError, SigningAlg},
+    p1363::der_to_p1363,
+    raw_signature::{
+        openssl::{cert_chain::check_chain_order, OpenSslMutex},
+        RawSigner, RawSignerError, SigningAlg,
+    },
     time_stamp::TimeStampProvider,
 };
 
-/// Implements `RawSigner` trait using OpenSSL's implementation of
-/// Edwards Curve encryption.
-pub struct Ed25519Signer {
+enum EcdsaSigningAlg {
+    Es256,
+    Es384,
+    Es512,
+}
+
+/// Implements `Signer` trait using OpenSSL's implementation of
+/// ECDSA encryption.
+pub struct EcdsaSigner {
+    alg: EcdsaSigningAlg,
+
     cert_chain: Vec<X509>,
     cert_chain_len: usize,
 
-    private_key: PKey<Private>,
+    private_key: EcKey<Private>,
 
     time_stamp_service_url: Option<String>,
     time_stamp_size: usize,
 }
 
-impl Ed25519Signer {
+impl EcdsaSigner {
     pub(crate) fn from_cert_chain_and_private_key(
         cert_chain: &[u8],
         private_key: &[u8],
+        alg: SigningAlg,
         time_stamp_service_url: Option<String>,
     ) -> Result<Self, RawSignerError> {
+        let alg = match alg {
+            SigningAlg::Es256 => EcdsaSigningAlg::Es256,
+            SigningAlg::Es384 => EcdsaSigningAlg::Es384,
+            SigningAlg::Es512 => EcdsaSigningAlg::Es512,
+            _ => {
+                return Err(RawSignerError::InternalError(
+                    "EcdsaSigner should be used only for SigningAlg::Es***".to_string(),
+                ));
+            }
+        };
+
         let _openssl = OpenSslMutex::acquire()?;
 
         let cert_chain = X509::stack_from_pem(cert_chain)?;
@@ -52,14 +77,13 @@ impl Ed25519Signer {
             ));
         }
 
-        let private_key = PKey::private_key_from_pem(private_key)?;
+        let private_key = EcKey::private_key_from_pem(private_key)?;
 
-        Ok(Ed25519Signer {
+        Ok(EcdsaSigner {
+            alg,
             cert_chain,
             cert_chain_len,
-
             private_key,
-
             time_stamp_service_url,
             time_stamp_size: 10000,
             // TO DO: Call out to time stamp service to get actual time stamp and use that size?
@@ -67,17 +91,30 @@ impl Ed25519Signer {
     }
 }
 
-impl RawSigner for Ed25519Signer {
+impl RawSigner for EcdsaSigner {
     fn sign(&self, data: &[u8]) -> Result<Vec<u8>, RawSignerError> {
         let _openssl = OpenSslMutex::acquire()?;
 
-        let mut signer = Signer::new_without_digest(&self.private_key)?;
+        let private_key = PKey::from_ec_key(self.private_key.clone())?;
 
-        Ok(signer.sign_oneshot_to_vec(data)?)
+        let mut signer = match self.alg {
+            EcdsaSigningAlg::Es256 => Signer::new(MessageDigest::sha256(), &private_key)?,
+            EcdsaSigningAlg::Es384 => Signer::new(MessageDigest::sha384(), &private_key)?,
+            EcdsaSigningAlg::Es512 => Signer::new(MessageDigest::sha512(), &private_key)?,
+        };
+
+        signer.update(data)?;
+
+        let der_sig = signer.sign_to_vec()?;
+        der_to_p1363(&der_sig, self.alg())
     }
 
     fn alg(&self) -> SigningAlg {
-        SigningAlg::Ed25519
+        match self.alg {
+            EcdsaSigningAlg::Es256 => SigningAlg::Es256,
+            EcdsaSigningAlg::Es384 => SigningAlg::Es384,
+            EcdsaSigningAlg::Es512 => SigningAlg::Es512,
+        }
     }
 
     fn reserve_size(&self) -> usize {
@@ -94,7 +131,7 @@ impl RawSigner for Ed25519Signer {
     }
 }
 
-impl TimeStampProvider for Ed25519Signer {
+impl TimeStampProvider for EcdsaSigner {
     fn time_stamp_service_url(&self) -> Option<String> {
         self.time_stamp_service_url.clone()
     }
