+++ internal/crypto/src/raw_signature/webcrypto/check_certificate_trust.rs
@@ -20,8 +20,9 @@ use x509_parser::{
 use crate::{
     cose::{CertificateTrustError, CertificateTrustPolicy},
     p1363::der_to_p1363,
-    raw_signature::{RawSignatureValidationError, SigningAlg},
-    webcrypto::async_validator_for_signing_alg,
+    raw_signature::{
+        webcrypto::async_validator_for_signing_alg, RawSignatureValidationError, SigningAlg,
+    },
 };
 
 pub(crate) async fn check_certificate_trust(
