+++ sdk/src/error.rs
@@ -325,11 +325,15 @@ pub enum Error {
 pub type Result<T> = std::result::Result<T, Error>;
 
 #[cfg(target_arch = "wasm32")]
-impl From<c2pa_crypto::webcrypto::WasmCryptoError> for Error {
-    fn from(err: c2pa_crypto::webcrypto::WasmCryptoError) -> Self {
+impl From<c2pa_crypto::raw_signature::webcrypto::WasmCryptoError> for Error {
+    fn from(err: c2pa_crypto::raw_signature::webcrypto::WasmCryptoError) -> Self {
         match err {
-            c2pa_crypto::webcrypto::WasmCryptoError::UnknownContext => Self::WasmInvalidContext,
-            c2pa_crypto::webcrypto::WasmCryptoError::NoCryptoAvailable => Self::WasmNoCrypto,
+            c2pa_crypto::raw_signature::webcrypto::WasmCryptoError::UnknownContext => {
+                Self::WasmInvalidContext
+            }
+            c2pa_crypto::raw_signature::webcrypto::WasmCryptoError::NoCryptoAvailable => {
+                Self::WasmNoCrypto
+            }
         }
     }
 }
