+++ sdk/src/utils/test.rs
@@ -522,7 +522,7 @@ impl AsyncSigner for WebCryptoSigner {
     }
 
     async fn sign(&self, claim_bytes: Vec<u8>) -> crate::error::Result<Vec<u8>> {
-        use c2pa_crypto::webcrypto::WindowOrWorker;
+        use c2pa_crypto::raw_signature::webcrypto::WindowOrWorker;
         use js_sys::{Array, Object, Reflect, Uint8Array};
         use wasm_bindgen_futures::JsFuture;
         use web_sys::CryptoKey;
